# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Terrace Farming System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /readings:
    get:
      operationId: getReadings
      tags: [readings]
      summary: List sensor readings
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of readings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SensorReading"
  /readings/latest:
    get:
      operationId: getLatestReading
      tags: [readings]
      summary: Get latest sensor reading
      responses:
        "200":
          description: Latest reading
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorReading"
        "404":
          description: No readings found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /readings/manual:
    post:
      operationId: createManualReading
      tags: [readings]
      summary: Create a manual sensor reading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadingBody"
      responses:
        "201":
          description: Reading created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorReading"
...
  /readings/device:
    post:
      operationId: createDeviceReading
      tags: [readings]
      summary: Receive reading from IoT device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceReadingBody"
      responses:
        "201":
          description: Reading received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensorReading"
...
  /prices:
    get:
      operationId: getPrices
      tags: [prices]
      summary: List price entries
      parameters:
        - name: cropName
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of price entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceEntry"
    post:
      operationId: createPriceEntry
      tags: [prices]
      summary: Add a price entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePriceEntryBody"
      responses:
        "201":
          description: Price entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceEntry"
...
  /prices/prediction/{cropName}:
    get:
      operationId: getPricePrediction
      tags: [prices]
      summary: Get price prediction for a crop
      parameters:
        - name: cropName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Price prediction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricePrediction"
        "404":
          description: No data for this crop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /prices/{id}:
    delete:
      operationId: deletePriceEntry
      tags: [prices]
      summary: Delete a price entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SensorReading:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        farmId:
          type: ["integer", "null"]
        deviceId:
          type: ["string", "null"]
        soilMoisture:
          type: number
        temperature:
          type: number
        humidity:
          type: number
        lightIntensity:
          type: number
        waterLevel:
          type: number
        sourceType:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - soilMoisture
        - temperature
        - humidity
        - lightIntensity
        - waterLevel
        - sourceType
        - createdAt
    ReadingBody:
      type: object
      properties:
        soilMoisture:
          type: number
        temperature:
          type: number
        humidity:
          type: number
        lightIntensity:
          type: number
        waterLevel:
          type: number
      required:
        - soilMoisture
        - temperature
        - humidity
        - lightIntensity
        - waterLevel
    DeviceReadingBody:
      type: object
      properties:
        deviceId:
          type: string
        soilMoisture:
          type: number
        temperature:
          type: number
        humidity:
          type: number
        lightIntensity:
          type: number
        waterLevel:
          type: number
      required:
        - deviceId
        - soilMoisture
        - temperature
        - humidity
        - lightIntensity
        - waterLevel
...
    PriceEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        cropName:
          type: string
        pricePerKg:
          type: number
        marketName:
          type: string
        unit:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - cropName
        - pricePerKg
        - marketName
        - unit
        - createdAt
    CreatePriceEntryBody:
      type: object
      properties:
        cropName:
          type: string
        pricePerKg:
          type: number
        marketName:
          type: string
        unit:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - cropName
        - pricePerKg
    PricePrediction:
      type: object
      properties:
        cropName:
          type: string
        currentPrice:
          type: number
        nextDayPrice:
          type: number
        nextWeekPrice:
          type: number
        trend:
          type: string
        trendPercent:
          type: number
        dataPoints:
          type: integer
        note:
          type: string
      required:
        - cropName
        - currentPrice
        - nextDayPrice
        - nextWeekPrice
        - trend
        - trendPercent
        - dataPoints
        - note
...