# lib/api-zod/src/generated/types/pricePrediction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Terrace Farming System API
 * OpenAPI spec version: 0.1.0
 */

export interface PricePrediction {
  cropName: string;
  currentPrice: number;
  nextDayPrice: number;
  nextWeekPrice: number;
  trend: string;
  trendPercent: number;
  dataPoints: number;
  note: string;
}